Model Formulation

The model code is divided into four sections. Each code file sets up a portion of the final model. Description of all models and variables in the model can be found below.

setup_base.jl

setup_storage.jl

setup_transport.jl

setup_policy.jl

setup_mass_balance.jl

H2CS2.setup_base — Function

setup_base(model::Model, inputs::InputStruct)

Establish an economic dispatch problem with all components except the mass balance constraint.

The returned model has the following formulation:

\[\begin{aligned} \min \quad &\sum_{i \in P} \sum_{y \in Y } \delta_y C_{i,y}^I a_{i,y}+ \sum_{i \in P}\sum_{y \in Y} \delta_y C_{i,y}^F (b_{i,y} + a_{i,y} - r_{i,y}) +\\ &\sum_{i \in P}\sum_{y \in Y} \sum_{h \in H} \delta_y w_h C^V_{i,y,h} q_{i,y,h} \end{aligned}\]

\[\text{subject to}\]

Producer availability:

\[ 0 \leq q_{i,y,h} \leq \Gamma_{i,y,h} (b_{i,y} + a_{i,y} - r_{i,y}) \qquad\forall i, y , h\]

Added capacity restrictions:

\[ 0 \leq a_{i,y} \leq A_{i,y}^{max} \qquad\forall i, y\]

Retired capacity restrictions:

\[ 0\leq r_{i,y} \leq b_{i,y} \qquad\forall i,y\]

Annual capacity transfers:

\[ b_{i,y} = b_{i,y-1} + a_{i,y-1} -r_{i,y-1} \qquad\forall y>0, i\]

Initial Capacity:

\[ b_{i,0} = B_{i,0} \qquad\forall i\]

H2CS2.setup_storage — Function

setup_storage(model::Model, inputs::InputStruct)

This function modifies the model created by "setup_base" in order to add storage units to the model.

These units store energy within a given simulation year. The storage levels at the beginning of the year are optimized by the model. Each storage resevoir may thus begin the year either full, empty, or anywhere in between. At the end of each year, the storge unit must return to the same level as at the beginning.

Storage units have two different types of capacity. The first, henceforth refered to as charge/discharge capacity, indicates the rate at which hydrogen can be added or removed from the resevoir. This capacity is analagous to the "power capacity" of battery storage systems. In our model, we determine this capacity endogenously, meaning that the model can choose how much to build or retire. The second capacity, henceforth called the storage capacity, indicates the total quantity of hydrogen that can be held in a given storage unit. This is analagous to the "energy capacity" of battery storage systems. In our model, the storage capacity is an exogenous input parameter.

To implement storage, we make the following modifications to the model:

\[\begin{aligned} Objective \mathrel{+}=&\sum_{j \in S} \sum_{y \in Y } \delta_y C_{j,y}^I a_{j,y}+ \sum_{j \in S}\sum_{y \in Y} \delta_y C_{j,y}^F (b_{j,y} + a_{j,y} - r_{j,y}) +\\ &\sum_{j \in S}\sum_{y \in Y} \sum_{h \in H} \delta_y w_h C^V_{j,y,h} q_{i,y,h}^c \end{aligned}\]

Storage charge availability

\[0 \leq q_{j,y,h}^c \leq \Gamma_{j,y,h} (b_{j,y} + a_{j,y} - r_{j,y}) \qquad\forall j, y , h\]

Storage discharge availability

\[0 \leq q_{j,y,h}^d \leq \Gamma_{j,y,h} (b_{j,y} + a_{j,y} - r_{j,y}) \qquad\forall j, y , h\]

Added charge/discharge capacity restrictions

\[0 \leq a_{j,y} \leq A_{j,y}^{max} \qquad\forall j, y\]

Retired capacity restrictions

\[0\leq r_{j,y} \leq b_{j,y} \qquad\forall j,y\]

Annual capacity transfers

\[b_{j,y} = b_{j,y-1} + a_{j,y-1} -r_{j,y-1} \qquad\forall y>0, j\]

Initial charge/discharge capacity

\[b_{j,0} = B_{j,0} \qquad\forall i\]

Maximum storage capacity

\[0\leq s_{j,y,h} \leq S_{j}^{max} \qquad \forall j\]

Storage levels after charging and discharging

\[s_{j,y,h} = (1- \lambda_{j})^{w_{h}}s_{j,y-1,h} + \eta_j^c w_h q^{c}_{j,y,h} - \frac{q_{j,y,h}^d w_h}{\eta_j^{d}} \qquad \forall j,y,h\]

The last constraint is set to be periodic, so that the state of charge at the end of each year is the same as at the beginning of the year.

H2CS2.setup_transport — Function

setup_transport(model::Model, inputs::InputStruct)

This function modifies the base model by adding regional transportation.

The following equations are added to the model:

\[ Objective \mathrel{+}= \sum_{e \in E} \sum_{y \in Y}\sum_{h \in H} \delta_y w_h C^{Trans}_{e,y} f_{e,y,h}\]

Flow limits:

\[ 0 \leq f_{r,j,y,h} \leq F_{r,j,y}^{max}\qquad \forall r,j,y,h\]

H2CS2.setup_policy — Function

setup_policy(model::Model, inputs::InputStruct)

This function modifies the base model by adding a carbon tax.

The carbon tax is implemented as a variable production cost in the objective function. In the future, this function could be expanded to allow for more complex policies.

\[ Objective \mathrel{+}= \sum_{i \in P}\sum_{y \in Y} \sum_{h \in H} \delta_y w_h \tau_{CO_2} q_{i,y,h} \epsilon_{i}\]

H2CS2.setup_mass_balance — Function

setup_mass_balance(model::Model, inputs::InputStruct)

This function modifies the model by adding regional mass balance constraints

The added constraint is shown below. In the constraint formulation, $P_r$, $S_r$, and $K_r$ represent the producers, storage units, and consumers that are located in region $r$.

Mass balance:

\[\begin{aligned} 0 =&\sum_{i\in P_r} q_{i,y,h} + \sum_{j\in S_r} (q_{j,y,h}^d-q_{j,y,h}^c)\\ & - \sum_{k \in K_r} D_{i,h,r} -\sum_{\substack{e \in E\\ \text{s.t}\: e = (r,j)}} (f_{e,y,h}) + \sum_{\substack{e \in E\\ \text{s.t}\: e = (j,r)} } (f_{e,y,h}) \qquad\forall r, y, h \end{aligned}\]