Model Coefficients

Each model coefficient is defined through a function that reads the coefficient value from the input data. All coefficient functions have names beginning with capital letters. The code for all of the coefficients is contained in the coefficients.jl file.

In this page, we proviide documentation for each of the coefficient functions. The documentation for each coefficient specifies exactly how the coefficient values are calculated from the input data.

H2CS2.Cost_VOM — Function

Cost_VOM(inputs::InputStruct, idx_prod::Int, idx_year::Int, idx_hour::Int)

Extracts the variable cost ($C_{i,y,h}^V$) of a specified producer in a given year and hour.

For each producer, the variable cost is calculated as:

\[\begin{aligned} C_{i,y,h}^V =\:& \mathrm{nonfuel\_variable\_cost}\:+\\ & \mathrm{electricity\_requirement}*\mathrm{electricity\_price}\:+\\ & \mathrm{gas\_requirement}*\mathrm{gas\_price} \end{aligned}\]

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_producer (Int)... index of producer for which to get cost

idx_year (Int) ... index of year for which to get cost

idx_hour (Int)... index of hour for which to get cost

H2CS2.Cost_VOM_Storage — Function

Cost_VOM_Storage(inputs::InputStruct, idx_stor::Int, idx_year::Int, idx_hour::Int)

Extracts the variable cost ($C_{j,y,h}^V$) of a specified storage unit in a given year and hour.

For each storage unit, the variable cost is calculated as:

\[\begin{aligned} C_{j,y,h}^V =\:& \mathrm{nonfuel\_variable\_cost}\:+\\ & \mathrm{electricity\_requirement}*\mathrm{electricity\_price}\:+\\ & \mathrm{gas\_requirement}*\mathrm{gas\_price} \end{aligned}\]

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

idx_year (Int) ... index of year for which to get cost

idx_hour (Int)... index of hour for which to get cost

H2CS2.Cost_FOM — Function

Cost_FOM(inputs::InputStruct, idx_prod::Int, idx_year::Int)

Extracts the annual fixed cost cost ($C_{i,y,h}^F$) of a specified producer in a given year.

For each producer, the fixed cost is taken directly from the producer inputs. We currently assume that annual fixed costs do not change over time.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_producer (Int)... index of producer for which to get cost

idx_year (Int) ... index of year for which to get cost

H2CS2.Cost_FOM_Storage — Function

Cost_FOM_Storage(inputs::InputStruct, idx_stor::Int, idx_year::Int)

Extracts the annual fixed cost cost ($C_{j,y,h}^F$) of a specified storage unit in a given year.

For each storage unit, the fixed cost is taken directly from the storage inputs. We currently assume that annual fixed costs do not change over time.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

idx_year (Int) ... index of year for which to get cost

H2CS2.Cost_Invest — Function

Cost_Invest(inputs::InputStruct, idx_prod::Int, idx_year::Int)

Extracts the investment cost ($C_{i,y,h}^I$) of a specified producer in a given year.

The investment cost is only applied during the year in which a producer adds capacity. For all other years, the investment cost is zero. When applicable, the investment cost is taken directly from the producer inputs.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_producer (Int)... index of producer for which to get cost

idx_year (Int) ... index of year for which to get cost

H2CS2.Cost_Invest_Storage — Function

Cost_Invest_Storage(inputs::InputStruct, idx_prod::Int, idx_year::Int)

Extracts the investment cost ($C_{j,y,h}^I$) of a specified storage unit in a given year.

The investment cost is only applied during the year in which a storage unit adds capacity. For all other years, the investment cost is zero. When applicable, the investment cost is taken directly from the storage inputs.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

idx_year (Int) ... index of year for which to get cost

H2CS2.Producer_Availability — Function

Producer_Availability(inputs::InputStruct, idx_prod::Int, idx_year::Int, idx_hour::Int)

Extracts the producer availability factor ($\Gamma_{i,y,h}$) of a specified producer in a given year and hour.

For each producer, the availability factor is calculated as:

\[ \Gamma_{i,y,h} =\: \mathrm{base\_availability}*\mathrm{scale\_factor}\]

where the base availability is taken directly from the "producer" inputs and the scale factor is taken for the corresponding column in the "time" inputs. The scale factor varies by year and hour.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_producer (Int)... index of producer for which to get cost

idx_year (Int) ... index of year for which to get cost

idx_hour (Int)... index of hour for which to get cost

H2CS2.Max_Producer_Capacity — Function

Max_Producer_Capacity(inputs::InputStruct, idx_prod::Int, idx_year::Int)

Extracts the maximum capacity ($A_{i,y}^{max}$) that can be added by a producer in a given year.

By design, each producer can only build capacity in a single year. This "build year" is specified by the "year" column of the producer inputs. During that year, the max buildable capacity is taken directly from the producer inputs. During every other year, the maximum buildable capacity is set to zero.

Capacity expansion in mulitple years can be incorporated by creating a new producer for each simulation year. For instance, once could create a "electrolyzer2030", which can be built in 2030 and an "electrolyzer2040", which can be built in 2040. This would allow the model to build capacity in both years.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_producer (Int)... index of producer for which to get cost

idx_year (Int) ... index of year for which to get cost

H2CS2.Max_Storage_Capacity — Function

Max_Storage_Capacity(inputs::InputStruct, idx_prod::Int, idx_year::Int)

Extracts the maximum capacity ($A_{j,y}^{max}$) that can be added by a storage unit in a given year.

By design, each storage unit can only build capacity in a single year. This "build year" is specified by the "year" column of the storage inputs. During that year, the max buildable capacity is taken directly from the storage inputs. During every other year, the maximum buildable capacity is set to zero.

Capacity expansion in mulitple years can be incorporated by creating a new storage unit for each simulation year. For instance, once could create a "saltcaverns2030", which can be built in 2030 and an "saltcaverns2040", which can be built in 2040. This would allow the model to build capacity in both years.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

idx_year (Int) ... index of year for which to get cost

H2CS2.Max_Storage_Quantity — Function

Max_Storage_Quantitity(inputs::InputStruct, idx_stor::Int)

Extracts the maximum storage quantity ($S_{j}^{max}$) for each storage unit.

The storage limit is taken direclty from the "storage" input table.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

H2CS2.Existing_Capacity — Function

Existing_Capacity(inputs::InputStruct, idx_prod::Int)

Extracts the pre-existing capacity ($B_{i,0}$) for each producer at the beginning of the first simulation year.

The pre-existing capacity is taken direclty from the "producer" input table.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_prod (Int)... index of producer for which to get cost

H2CS2.Existing_Storage_Capacity — Function

Existing_Storage_Capacity(inputs::InputStruct, idx_stor::Int)

Extracts the pre-existing capacity ($B_{j,0}$) for each storage unit at the beginning of the first simulation year.

The pre-existing capacity is taken direclty from the "storage" input table.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

H2CS2.Flow_Limit — Function

Flow_Limit(inputs::InputStruct, idx_edge::Int, idx_hour::Int)

Extracts the pre-existing capacity ($F_{e,y}^{max}$) for a given transportation route and in a given year.

The flow limit is taken directly from the "transportation" input table. Note that each edge on the graph is active for only a single year. Hence, the flow limit function does not depend on the year.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_edge (Int)... index of edge for which to get the flow limit

idx_hour (Int) ... index of hour for which to get the flow limit

H2CS2.Cost_Transport — Function

Cost_Transport(inputs::InputStruct, idx_edge::Int, idx_year::Int)

Extracts the transportation cost factor ($C_{e,y}^{Trans}$) of along a specified transportation route (edge) for a given year.

For each edge, the cost is calculated as:

\[ C_{e,y}^{Trans} =\: \mathrm{distance}*\mathrm{cost\_per\_unit\_distance}\]

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_edge (Int)... index of transportation route for which to get cost.

idx_year (Int) ... index of year for which to get cost.

H2CS2.Discharge_Efficiency — Function

Discharge_Efficiency(inputs::InputStruct, idx_stor::Int)

Extracts the discharge efficiency ($\eta^d_{j}$) for each storage unit.

The discharge efficiency is taken directly from the "storage" input table.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

H2CS2.Charge_Efficiency — Function

Charge_Efficiency(inputs::InputStruct, idx_stor::Int)

Extracts the charge efficiency ($\eta^c_{j}$) for each storage unit.

The charge efficiency is taken directly from the "storage" input table.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

H2CS2.Self_Discharge_Rate — Function

Self_Discharge_Rate(inputs::InputStruct, idx_stor::Int)

Extracts the self-discharge rate ($\lambda_{j}$) for each storage unit.

The self-discharge rate is taken direclty from the "storage" input table.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_stor (Int)... index of storage unit for which to get cost

H2CS2.Weight_Hour — Function

Weight_Hour(inputs::InputStruct, idx_year::Int, idx_hour::Int)

Extracts the hour weight ($w_h$) of a given hour in a given year.

Each hour in the model is a "representative" for a collection of hours throughout the year. For instance, we may simulate only one hour to represent all hours witin the month of January. The hour weight gives the number of real-life hours represented by a the given simuation hour. It is used to weight the objective function. The hour weights are taken directly from the "time" input table.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_year (Int) ... index of year for which to get cost

idx_hour (Int)... index of hour for which to get cost

H2CS2.Discount_Factor — Function

Discount_Factor(inputs::InputStruct, idx_year::Int)

Calculates the discount factor $\delta_y$ for a given year.

The discount factor is currently taken as a single-year discount factor of the following form:

\[\delta_y = \frac{1}{(1 + r)^{t-t_0}}\]

where $r$ is the discount rate, $t$ is the current simulation year, and $t_0$ is the base simulation year.

ARGUMENTS:

inputs (InputStruct) ... data structure containing inputs

idx_year (Int) ... index of year for which to get cost